change enums from file to folder for  extendable feature
